Added basic support for form rendering
The default templates provided by SilverStripe are totally differents
from the approach taken by Bootstrap.

Implemented a subset of fields in a Bootstrap compatible way (for as
much as possible): rendering and PHP are not properly decoupled (e.g.
extraClasses is supposed to be appied to <input> but I found no way to
apply custom classes to the <div> holder).