Allow custom image URLs
Reapplied #0ab3b5c5dc996dee: colorbox by default detects if an URL is
an image by checking if the ending part is a common image extension
(.jpeg, .png etc...). This clashes with the custom URLs used by
silverstripe-adg to dynamically generate the technical drawing, so
that checking is now disabled.