Expect the favicon in the base URL by default
This correction is similar to what done by the commit 853df4021c57 to
the brand link.

Root URL and document base URL are not always the same, e.g. on my
development system the root URL is http://localhost/ while the document
base URL is http://localhost/mysite/. In this case the favicon is
expected to be found in http://localhost/mysite/favicon.ico.